requirements.txt
langchain
langgraph
langchain-core
langchain-community
fastapi
uvicorn
pydantic
python-dotenv
requests
langchain-openai
pytest
black
flake8
mypy